Harden debug and floating point exceptions
