+++ src/interrupts/handlers/exceptions/debug/types.rs
@@ -1,0 +1,26 @@
+// NONOS Operating System
+// Copyright (C) 2026 NONOS Contributors
+//
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU Affero General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+// GNU Affero General Public License for more details.
+//
+// You should have received a copy of the GNU Affero General Public License
+// along with this program. If not, see <https://www.gnu.org/licenses/>.
+
+#[derive(Debug, Clone, Copy)]
+pub struct DebugInfo {
+    pub dr6: u64,
+    pub breakpoint_0: bool,
+    pub breakpoint_1: bool,
+    pub breakpoint_2: bool,
+    pub breakpoint_3: bool,
+    pub single_step: bool,
+    pub task_switch: bool,
+}
